+++ src/lib.rs
@@ -9,7 +9,7 @@ pub mod message;
 
 pub use crate::class::{MidiClass, MidiReadError};
 pub use crate::packet::cable_number::CableNumber;
-pub use crate::packet::reader::MidiPacketBufferReader;
+pub use crate::packet::reader::UsbMidiPacketReader;
 pub use crate::packet::{MidiPacketParsingError, UsbMidiEventPacket};
 
 #[cfg(feature = "message-types")]